:hammer: Suggested changes
