+++ hassio/src/addon-store/hassio-addon-repository.ts
@@ -7,7 +7,7 @@ import {
   CSSResultArray,
 } from "lit-element";
 import "@polymer/paper-card/paper-card";
-import * as Fuse from "fuse.js";
+import memoizeOne from "memoize-one";
 
 import "../components/hassio-card-content";
 import { hassioStyle } from "../resources/hassio-style";
@@ -17,6 +17,7 @@ import {
   HassioAddonRepository,
 } from "../../../src/data/hassio";
 import { navigate } from "../../../src/common/navigate";
+import { filterAndSort } from "../components/hassio-filter-and-sort";
 
 class HassioAddonRepositoryEl extends LitElement {
   @property() public hass!: HomeAssistant;
@@ -26,45 +27,48 @@ class HassioAddonRepositoryEl extends LitElement {
 
   protected render(): TemplateResult | void {
     const repo = this.repo;
-    const addons = this.fuzzySearchAndSort(this.addons, this.filter);
+    const addons = memoizeOne(filterAndSort(this.addons, this.filter));
 
     if (this.filter && addons.length < 1) {
-      return html``;
-    } else {
       return html`
-        <div class="card-group">
-          <div class="title">
-            ${repo.name}
-            <div class="description">
-              Maintained by ${repo.maintainer}<br />
-              <a class="repo" href=${repo.url} target="_blank">${repo.url}</a>
-            </div>
-          </div>
-
-          ${addons.map(
-            (addon) => html`
-              <paper-card
-                .addon=${addon}
-                class=${addon.available ? "" : "not_available"}
-                @click=${this.addonTapped}
-              >
-                <div class="card-content">
-                  <hassio-card-content
-                    .hass=${this.hass}
-                    .title=${addon.name}
-                    .description=${addon.description}
-                    .available=${addon.available}
-                    .icon=${this.computeIcon(addon)}
-                    .iconTitle=${this.computeIconTitle(addon)}
-                    .iconClass=${this.computeIconClass(addon)}
-                  ></hassio-card-content>
-                </div>
-              </paper-card>
-            `
-          )}
+        <div class="description">
+          No results found in ${repo.name}
         </div>
       `;
     }
+    return html`
+      <div class="card-group">
+        <div class="title">
+          ${repo.name}
+          <div class="description">
+            Maintained by ${repo.maintainer}<br />
+            <a class="repo" href=${repo.url} target="_blank">${repo.url}</a>
+          </div>
+        </div>
+
+        ${addons.map(
+          (addon) => html`
+            <paper-card
+              .addon=${addon}
+              class=${addon.available ? "" : "not_available"}
+              @click=${this.addonTapped}
+            >
+              <div class="card-content">
+                <hassio-card-content
+                  .hass=${this.hass}
+                  .title=${addon.name}
+                  .description=${addon.description}
+                  .available=${addon.available}
+                  .icon=${this.computeIcon(addon)}
+                  .iconTitle=${this.computeIconTitle(addon)}
+                  .iconClass=${this.computeIconClass(addon)}
+                ></hassio-card-content>
+              </div>
+            </paper-card>
+          `
+        )}
+      </div>
+    `;
   }
 
   private computeIcon(addon) {
@@ -93,23 +97,6 @@ class HassioAddonRepositoryEl extends LitElement {
 
   private addonTapped(ev) {
     navigate(this, `/hassio/addon/${ev.currentTarget.addon.slug}`);
-  }
-
-  private fuzzySearchAndSort(addons, filter) {
-    if (!filter) {
-      return addons.sort((a, b) =>
-        a.name.toUpperCase() < b.name.toUpperCase() ? -1 : 1
-      );
-    }
-
-    const options: Fuse.FuseOptions<HassioAddonInfo> = {
-      keys: ["name", "description", "slug"],
-      caseSensitive: false,
-      minMatchCharLength: 2,
-      threshold: 0.2,
-    };
-    const fuse = new Fuse(addons, options);
-    return fuse.search(filter);
   }
 
   static get styles(): CSSResultArray {
